consolidate error classes
